fixDB migration: Update ApplicationDbContextModelSnapshot and add fixDB migration
This commit updates the ApplicationDbContextModelSnapshot class to set the "ProductVersion" annotation to "8.0.1" instead of "8.0.0". It also adds a new migration called fixDB, which is responsible for updating the database schema.

Additionally, this commit modifies the DishesDbModel class by adding the "required" attribute to the NameOfDish and Owner properties, indicating that these fields are mandatory.

Lastly, it adds a reference to the Newtonsoft.Json package in the UniEat.csproj file with version 13.0.3.